% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-body-ui.R
\docType{data}
\name{dashboard_body}
\alias{dashboard_body}
\title{Generated the dashboard main body}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
dashboard_body
}
\description{
this shinydashboard object contains all the tabItems of the app
and uses a specific shiny module for the scatterplot interface and menus.
}
\keyword{datasets}
