% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-header-ui.R
\docType{data}
\name{dashboard_header}
\alias{dashboard_header}
\title{Generate the shinydashboard header component}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
dashboard_header
}
\description{
Generate the shinydashboard header component
}
\keyword{datasets}
