% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-sidebar-ui.R
\docType{data}
\name{dashboard_sidebar}
\alias{dashboard_sidebar}
\title{left sidebar object with all menus}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
dashboard_sidebar
}
\description{
This inlcudes dataset selection
}
\keyword{datasets}
