% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms2dd.R
\name{dms2dd}
\alias{dms2dd}
\title{dms2dd}
\usage{
dms2dd(x = NULL)
}
\arguments{
\item{x}{location as "120°37′10″E"}
}
\value{
location as "120.619444444444"
}
\description{
convert geo location from degree minute seconds to
}
\examples{
dms2dd("120°37′10″E")
}
