% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.prob.of.being.in.R
\name{find.times.distribution}
\alias{find.times.distribution}
\title{extracts times of arrival and departure to/from spatial extent}
\usage{
find.times.distribution(Result, Spatial.Index)
}
\arguments{
\item{Result}{FLightR result object obtained from \code{\link{run.particle.filter}}}

\item{Spatial.Index}{Row numbers for spatial grid (\code{Result$Spatial$Grid}) to estimate schedules for.}
}
\value{
dataframe with columns for 0.025, 0.25, 0.5, 0.75, 0.975 probability of line crossing and rows for every crossing.
}
\description{
Idea of this functions is to extract schedules for known location
}
\author{
Eldar Rakhimberdiev
}

