% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightr.R
\name{flightR}
\alias{flightR}
\title{launches the flightR shiny app}
\usage{
flightR(ae.api.key = NULL, osn.login = NULL, osn.password = NULL,
  ...)
}
\arguments{
\item{ae.api.key}{a valid Aviation Edge API Key}

\item{osn.login}{a valid OpenSky Network login}

\item{...}{}

\item{osn.login}{a valid OpenSky Network password}
}
\value{
shiny application object
}
\description{
This function launches a shiny app which enables exploration of aviation
data collected by the flightR package. (to be continued with details of 
features)
}
\examples{
flightR::flightR()
}
