% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{FLightR2Movebank}
\alias{FLightR2Movebank}
\title{Summary of estimated locations for Movebank}
\usage{
FLightR2Movebank(Result, alpha = 0.5, filename = NULL)
}
\arguments{
\item{Result}{FLightR result object obtained from \code{\link{run.particle.filter}}}

\item{alpha}{coverage of the credible intervals for now only two options: 0.95 or 0.5.}

\item{filename}{if NULL data.frame in TAGS format will be returned otherwise .csv file in TAGS format will be written}
}
\value{
\code{NULL} if \code{filename} is provided or Movebank formatted dataframe.
}
\description{
Summarize result object in the format required for upload to Movebank
}
\details{
This function accepts FLightR results object.
}
\author{
Eldar Rakhimberdiev
}
