% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightr-server.R
\name{flightRserver}
\alias{flightRserver}
\title{Shiny app server function}
\usage{
flightRserver(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
details to be written here after
}
