% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightr-ui.R
\docType{data}
\name{flightRui}
\alias{flightRui}
\title{flightR shinydashboard UI}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
flightRui
}
\description{
using shinydashboard layout do define the main smartexplorer interface
}
\keyword{datasets}
