% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_ZI_distances}
\alias{get_ZI_distances}
\title{Estimate distances moved between twilights}
\usage{
get_ZI_distances(Result)
}
\arguments{
\item{Result}{An object created by \code{\link{run.particle.filter}}.}
}
\value{
a data frame containing median and quartiles for the distances and also departure and arrival time
}
\description{
This function estimate distances with all zeros from stationary periods. This means many of the resulting movements will have 0 as the distance
}
\author{
Eldar Rakhimberdiev
}
