% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ae-aircrafts.R
\name{getAEaircrafts}
\alias{getAEaircrafts}
\title{getAEaircrafts}
\usage{
getAEaircrafts(aircraft.registration.number = NULL,
  aircraft.hex.icao = NULL, airline.iata = NULL, api.key)
}
\arguments{
\item{aircraft.registration.number}{aircraft registration number}

\item{aircraft.hex.icao}{aircrafts hex ICAO code}

\item{airline.iata}{airline IATA code}

\item{api.key}{you authentication api key}
}
\value{
answer a json object containing airports information details
\itemize{
        \item "airplaneId": "55",
        \item "numberRegistration": "HB-JVE", 
        \item "productionLine": "Fokker 28/70/100", 
        \item "airplaneIataType": "F28 MK0100", 
        \item "planeModel": "F-100", 
        \item "modelCode": "F100", 
        \item "hexIcaoAirplane": "4B19EA", 
        \item "codeIataPlaneShort": "100", 
        \item "codeIataPlaneLong": "F100", 
        \item "constructionNumber": "11459", 
        \item "numberTestRgistration": "PH-EZD", 
        \item "rolloutDate": "0000-00-00", 
        \item "firstFlight": "1993-05-31T22:00:00.000Z", 
        \item "deliveryDate": "1993-06-29T22:00:00.000Z", 
        \item "registrationDate": "2004-05-11T22:00:00.000Z", 
        \item "lineNumber": "", "planeSeries": "", 
        \item "codeIataAirline": "2L", 
        \item "codeIcaoAirline": "", 
        \item "planeOwner": "", 
        \item "enginesCount": "2", 
        \item "enginesType": "JET", 
        \item "planeAge": "24", 
        \item "planeStatus": "active", 
      }
}
\description{
A function to get airport information from Aviation Edge API. As stated on 
aviation-edge.com, this api gives access to an aircraft registry in the form
of a database of all aircraft types

\strong{Available API Options}

\itemize{
   \item The entire database of aircrafts, i.e. get all aircrafts in one shot,
   this represents ~19k+ items (with free API)
   \item Information for a specific aircraft registration number.
   \item Information for aircrafts based on their hex ICAO code.
   \item All aircrafts of a specific airline, you can search based on airline IATA code.
}
}
\examples{

All aicrafts
\dontrun{ 
getAEaircrafts(
              api.key = 'key.here') 
}
Information about a specific airplane by registration number
\dontrun{ 
getAEaircrafts(
              aircraft.registration.number = 'HB-JVE',
              api.key = 'key.here') 
}
Information about a specific airplane by hex ICAO
\dontrun{ 
getAEaircrafts(
              aircraft.hex.icao = '4B19EA',
              api.key = 'key.here') 
}
 information about airplanes of a specific airline, by  IATA code. 
\dontrun{ 
getAEaircrafts(
              airline.iata = '0B',
              api.key = 'key.here') 
}


}
\seealso{
https://www.iata.org/publications/Pages/code-search.aspx
}
