% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ae-airports.R
\name{getAEairports}
\alias{getAEairports}
\title{getAEairports}
\usage{
getAEairports(airport.iata.code = NULL, country.iso.code = NULL,
  api.key)
}
\arguments{
\item{airport.iata.code}{IATA code of the airport}

\item{country.iso.code}{ISO code of the country (eg. DE)}

\item{api.key}{you authentication api key}
}
\value{
answer a json object containing airports information details
}
\description{
A function to get airport information from Aviation Edge API
}
\details{
\strong{The entire database of airports}
The entire data base of airports can be downloaded

\strong{Information for a specific airport}
For a designated airport, the data can be downloaded using it's IATA Code.

\strong{All airports for a specific country}
Using country code (Iso), all airports from this country can be retrieved
}
\seealso{
https://www.iata.org/publications/Pages/code-search.aspx
}
