% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ae-cities.R
\name{getAEcities}
\alias{getAEcities}
\title{getAEcities}
\usage{
getAEcities(city.iata.code = NULL, country.iso.code = NULL, api.key)
}
\arguments{
\item{city.iata.code}{IATA code of the airport}

\item{country.iso.code}{ISO code of the country (eg. DE)}

\item{api.key}{you authentication api key}
}
\value{
answer a json object containing airports information details
\itemize{
        \item "cityId": "1",
        \item "nameCity": "Anaa",
        \item "codeIataCity": "AAA", 
        \item "codeIso2Country": "PF", 
        \item "latitudeCity": "-17.05", 
        \item "longitudeCity": "-145.41667", 
        \item "timezone": "Pacific/Tahiti", 
        \item "GMT": "-10", 
        \item "geonameId": "0" 
      }
}
\description{
A function to get cities information from Aviation Edge API. As stated on 
aviation-edge.com, this cities database contains all cities in the world 
with their aviation and travel related data. 

\strong{Available API Options}

\itemize{
   \item The entire database of cities, i.e. get all cities in one shot,
   this represents ~XXk+ items worldwide (with free API)
   \item Information for a specific cityt using it's IATA Code.
   \item All cities in a specific country using country code (Iso).
}
}
\examples{

Details about Anaa city
\dontrun{ 
getAEcities(city.iata.code = 'AAA ',
              country.iso.code = NULL,
              api.key = 'key.here') 
}

All cities in France
\dontrun{
getAEcities(city.iata.code = NULL,
              country.iso.code = 'FR',
              api.key = 'key.here')
}

All cities
\dontrun{
getAEcities(city.iata.code = NULL,
              country.iso.code = NULL,
              api.key = 'key.here')
}

}
\seealso{
https://www.iata.org/publications/Pages/code-search.aspx

https://aviation-edge.com/developers/
}
