% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ae-countries.R
\name{getAEcountries}
\alias{getAEcountries}
\title{getAEcountries}
\usage{
getAEcountries(country.iso.code = NULL, country.name = NULL, api.key)
}
\arguments{
\item{country.iso.code}{ISO code of the country (eg. DE)}

\item{country.name}{name of the country (eg. France)}

\item{api.key}{you authentication api key}
}
\value{
answer a json object containing airports information details
\itemize{
        \item "countryId": "1", 
        \item "nameCountry": "Andorra", 
        \item "codeIso2Country": "AD", 
        \item "codeIso3Country": "AND", 
        \item "numericIso": "20", "population": "84000", 
        \item "capital": "Andorra la Vella", 
        \item "continent": "EU", 
        \item "nameCurrency": "Euro", 
        \item "codeCurrency": "EUR", 
        \item "codeFips": "AN", 
        \item "phonePrefix": "376" 
      }
}
\description{
A function to get cities information from Aviation Edge API. As stated on 
aviation-edge.com, this countries database contains all countries and regions
in the world.

\strong{Available API Options}

\itemize{
   \item The entire database of countries, i.e. get all countries in one shot,
   this represents ~XXk+ items worldwide (with free API)
   \item Information for a specific country using it's ISO Code.
   \item Information for a specific country using it's name.
}
}
\examples{

Details about Andorra
\dontrun{ 
getAEcountries(country.iso.code = "AD",
              api.key = 'key.here') 
}
\dontrun{ 
getAEcountries(country.name = "Andorra",
              api.key = 'key.here') 
}

All countries
\dontrun{
getAEcountries(country.iso.code = NULL,
              country.name = NULL,
              api.key = 'key.here')
}

}
\seealso{
https://www.iata.org/publications/Pages/code-search.aspx

https://aviation-edge.com/developers/
}
