% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-state-vectors-osn.R
\name{getAllStateVectorsOSN}
\alias{getAllStateVectorsOSN}
\title{getAllStateVectorsOSN}
\usage{
getAllStateVectorsOSN(icao24 = NULL, time = NULL,
  latitude.min = NULL, longitude.min = NULL, latitude.max = NULL,
  longitude.max = NULL, login = NULL, password = NULL)
}
\arguments{
\item{icao24}{character vector, one or more ICAO24 transponder addresses (e.g. "abc9f3")}

\item{time}{character, to retrieve flights vector as "YYYY-MM-DD HH:MM:SS"}

\item{latitude.min}{numeric, lower bound for the latitude in decimal degrees (optional)}

\item{longitude.min}{numeric, lower bound for the longitude in decimal degrees}

\item{latitude.max}{numeric, upper bound for the latitude in decimal degrees}

\item{longitude.max}{numeric, upper bound for the longitude in decimal degrees}

\item{login}{OSN login information}

\item{password}{OSN password information associated to the login provided}
}
\value{
The response is a JSON array of flights where each flight is an object with properties
}
\description{
A function to get all available details of flights by plane icao transponder adress
at a specific time - call to Opensky Network API.
}
\details{
Limitations for anonymous (unauthenticated) users can only get the most recent
state vectors, i.e. the time parameter will be ignored.
Anonymous users can only retrieve data with a time resultion of 10 seconds.
That means, the API will return state vectors for time now−(now mod 10).
This annonymous mode is not implemented in this function

Limitations for OpenSky users
OpenSky users can retrieve data of up to 1 hour in the past.
If the time parameter has a value t<now−3600 the API will return 400 Bad Request.
OpenSky users can retrieve data with a time resultion of 5 seconds.
That means, if the time parameter was set to t, the API will return state vectors
for time t−(t mod 5).
}
\examples{
\dontrun{
getAllStateVectorsOSN(icao24 = NULL,
                      time = "2018-01-29 00:00:00",
                      login = "login",
                      password = "my pasword")
}
}
\seealso{
https://opensky-network.org/apidoc/rest.html#all-state-vectors

https://en.wikipedia.org/wiki/Aviation_transponder_interrogation_modes#ICAO_24-bit_address
}
