% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-flights-by-airports-osn.R
\name{getFlightsByAirportOSN}
\alias{getFlightsByAirportOSN}
\title{getFlightsByAirportOSN}
\usage{
getFlightsByAirportOSN(icao.code = "JFK", type = "arrival",
  from.time = "2018-12-24 20:00:00", to.time = "2018-12-24 23:59:00",
  login = NULL, password = NULL)
}
\arguments{
\item{icao.code}{ICAO code of the airport}

\item{type}{either arrival or departure}

\item{from.time}{Start of time interval to retrieve flights as "YYYY-MM-DD HH:MM:SS"}

\item{to.time}{End of time interval to retrieve flights as "YYYY-MM-DD HH:MM:SS"}

\item{login}{osn valid login}

\item{password}{osn valid password}
}
\value{
The response is a JSON array of flights where each flight is an object with properties
}
\description{
A function to get all arrival or departure flights in a specific airport in a given time
interval - call to Opensky Network API
}
\details{
Retrieve flights for a certain airport which arrived or departured within a given time 
interval (begin, end). If no flights are found for the given period, 
HTTP stats 404 - Not found is returned with an empty response body.
}
\seealso{
https://www.world-airport-codes.com/

https://opensky-network.org/apidoc/rest.html#arrivals-by-airport

https://opensky-network.org/apidoc/rest.html#departures-by-airport
}
