% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-flight-tracker-ae.R
\name{getFlightTrackerAE}
\alias{getFlightTrackerAE}
\title{getFlightTrackerAE}
\usage{
getFlightTrackerAE(depIata = NULL, depIcao = NULL, arrIata = NULL,
  arrIcao = NULL, aircraftIcao = NULL, regNum = NULL,
  aircraftIcao24 = NULL, airlineIata = NULL, airlineIcao = NULL,
  flightIata = NULL, flightIcao = NULL, flightNum = NULL,
  status = NULL, limit = NULL, api.key)
}
\arguments{
\item{depIata}{Departure airport IATAcode}

\item{depIcao}{Departure airport ICAO code}

\item{arrIata}{Arrival airport IATA code}

\item{arrIcao}{Arrival airport ICAO code}

\item{aircraftIcao}{Aircraft ICAO code}

\item{regNum}{Aircraft registration number}

\item{aircraftIcao24}{Aircraft ICAO24 code}

\item{airlineIata}{Airline IATA code}

\item{airlineIcao}{Airline ICAO code}

\item{flightIata}{Flight IATA code}

\item{flightIcao}{Flight ICAO code}

\item{flightNum}{Flight number}

\item{status}{Status of the flight (en-route, landed)}

\item{limit}{Limit value for output}

\item{api.key}{you authentication api key}
}
\value{
answer a json object containing flights list with status and details
}
\description{
A function to get information about all flights in the world with
}
\details{
\strong{Flight Tracker API}
Clients can connect to the flight tracking through the API key which gives 
direct access to all the live flight data. The flight tracker API allows 
clients to display within their applications or on their websites and other 
platforms, the actual tracking of flights including all their relevant 
information. This service is central to our aviation data and maintained
constantly within our database. Since the data is live, the updates happen 
within short intervals of just a few minutes. Aviation Edge’s function is 
to gather and maintain the complete aviation data for our clients so that 
they can create the functional tools for their customers.
 
\strong{Sources and process}
The data of the flight tracking API is collected through ADS-B systems as
well as the use of flight schedules to anticipate the location and fill in
the gaps where needed. We also receive data from our aviation partners and 
combine all this to the central system for our clients.

\strong{Data included in the Flight Tracker API}
Geography: Location information such as latitude, longitude, altitude and direction.
Speed: Vertical and horizontal speed of aircraft.
Departure and arrival: IATA codes and ICAO codes of the departure and arrival location.
Aircraft and flight: IATA and ICAO number of flight and registration number, ICAO code and ICAO24 code of aircraft.
Airline: IATA code and ICAO code of airline.
System information: Squawk, status and last updated.
}
\seealso{
https://www.iata.org/publications/Pages/code-search.aspx
}
