% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-osn-flights-by-aircraft.R
\name{getOSNflightsByAircraft}
\alias{getOSNflightsByAircraft}
\title{getOSNflightsByAircraft}
\usage{
getOSNflightsByAircraft(icao24 = NULL, from.time = NULL,
  to.time = NULL, login = NULL, password = NULL)
}
\arguments{
\item{icao24}{Unique ICAO 24-bit address of the transponder in hex string representation. All letters need to be lower case}

\item{login}{osn valid login}

\item{password}{osn valid password}

\item{begin.time}{as "YYYY-MM-DD HH:MM:SS", Start of time interval to retrieve flights}

\item{end.time}{as "YYYY-MM-DD HH:MM:SS", End of time interval to retrieve flights}
}
\value{
The response is a JSON array of flights where each flight is an object with properties:
}
\description{
Documentation from OpenSky Network API
This API call retrieves flights for a particular aircraft within a certain 
time interval. Resulting flights departed and arrived within (begin, end)
If no flights are found for the given period, HTTP stats 404 - Not found is
returned with an empty response body.
}
\seealso{
https://opensky-network.org/apidoc/rest.html#flights-by-aircraft
}
