% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{make.calibration}
\alias{make.calibration}
\title{Creates a calibration object, further used for calculation of coordinates in the \code{\link{run.particle.filter}}.}
\usage{
make.calibration(Proc.data, Calibration.periods, model.ageing = FALSE,
  plot.each = FALSE, plot.final = FALSE)
}
\arguments{
\item{Proc.data}{processed data object generated by \code{\link{get.tags.data}}
Calibration.periods a data frame containing start and end dates of all the calibration periods (POSIXct) and geographic coordinates of the corresponding calibration locations.}
}
\description{
Function estimates all nesessary parameters from the calibration data logged in a known location or locations.
}

