% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-timetable-item-overview-ae.R
\name{makeTimetableItemOverviewAE}
\alias{makeTimetableItemOverviewAE}
\title{makeTimetableItemOverviewAE}
\usage{
makeTimetableItemOverviewAE(timetable.item = NULL)
}
\arguments{
\item{flight}{a list containing the airport timetable item (one flight)}
}
\value{
a list with main details of that file
}
\description{
A function which takes an airport timetable item (see getFlightSchedulesTimetables)
and returnd a list containing the flight main metadata
}
\examples{
makeTimetableOverview(timetable)
}
\seealso{
\code{\link{getFlightSchedulesTimetables}}
}
