% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plotting_functions.R
\name{plot.likelihood}
\alias{plot.likelihood}
\title{plot likelihood surface over map}
\usage{
plot.likelihood(object, date = NULL, twilight.index = NULL)
}
\arguments{
\item{object}{either output from \code{\link{run.particle.filter}} or \code{\link{run.particle.filter}}}

\item{date}{either NULL or a date (possibly with time) closest to the twilight you wan tto be plotted}

\item{twilight.index}{number of likelohood surface to be plotted}
}
\description{
plots specific likelihood surface over map
}
\details{
function plots likelihoods before particle filter run, so these are pure results of calibrations without any movement model
}

