% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plotting_functions.R
\name{plot.lon.lat}
\alias{plot.lon.lat}
\title{plots result by longitude and latitude}
\usage{
plot.lon.lat(Result, scheme = c("vertical", "horizontal"))
}
\arguments{
\item{Result}{FLightR result object obtained from \code{\link{run.particle.filter}}}

\item{scheme}{either 'vertical' or 'horizontal' layouts
return NULL}
}
\description{
This function plots result by latitude and longitude in either vertical or horizontal layout
}

