% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plotting_functions.R
\name{plot.util.distr}
\alias{plot.util.distr}
\title{plots resulting track over map with uncertainty shown by spave utilisation distribution}
\usage{
plot.util.distr(Result, dates = NULL, map.options = NULL,
  percentiles = c(0.4, 0.6, 0.8), zoom = "auto",
  geom_polygon.options = NULL, save.options = NULL, color.palette = NULL,
  use.palette = TRUE, background = NULL, plot = TRUE, save = TRUE,
  add.scale.bar = FALSE, scalebar.options = NULL)
}
\arguments{
\item{Result}{FLightR result object obtained from \code{\link{run.particle.filter}}}

\item{dates}{Use NULL if all twilights will be used for plotting, one integer if specific twilight should be plotted (line number in Result$Results$Quantiles). Use data.frame with first colum - start of the period and second - end of the period and each line represents a new period to plot specific periods, e.g. wintering or migration.}

\item{map.options}{options passed to \code{\link[ggmap]{get_map}}, note that \code{zoom} option is defined separately}

\item{percentiles}{Probability breaks for utilisation distribution}

\item{zoom}{Zoom for map. If 'auto' FLightR will try to find optimal zoom level by downloading different size maps and checking whether all the points fit the map.}

\item{geom_polygon.options}{passed to \code{\link[ggplot2]{geom_polygon}}}

\item{save.options}{ptions passed to \code{\link[ggplot2]{ggsave}}. Filename should be defined here.}

\item{color.palette}{colors for probability contours. Either NULL or \code{\link[grDevices]{colorRampPalette}} object}

\item{use.palette}{should the same colors be used for polygon boundaries as for polygon filling?}

\item{background}{if provided will be used as a background. Must be created by \code{link[ggmap]{get_map}}}

\item{plot}{should function produce a plot?}

\item{save}{should function save results with \code{\link[ggplot2]{ggsave}}?}

\item{add.scale.bar}{will add scalebar with the \code{\link[ggsn]{scalebar}}}

\item{scalebar.options}{options passed to \code{\link[ggsn]{scalebar}}}
}
\value{
list with two parts 
        \item{res_buffers}{spatial buffers for defined probability values}
        \item{p}{\code{\link[ggplot2]{ggplot}} object}
}
\description{
May be use not only for the whole track but for a set of specific dates, e.g. to show spatial uncertainty during migration
}
\author{
Eldar Rakhimberdiev
}

