% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-null-by-na.R
\name{replaceNullByNA}
\alias{replaceNullByNA}
\title{replaceNullByNA}
\usage{
replaceNullByNA(x = NULL)
}
\arguments{
\item{x}{a list}
}
\value{
x a list
}
\description{
function to replace NULL Values in a list by NA values
}
\examples{
replaceNullByNA(list("ab", "bc", NULL, "de"))
}
