% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-wikipedia-airports-details.R
\name{scrapWikipediaAirportsDetails}
\alias{scrapWikipediaAirportsDetails}
\title{scrapWikipediaAirportsDetails}
\usage{
scrapWikipediaAirportsDetails(url)
}
\arguments{
\item{url}{an airport wikipedia ura}
}
\value{
list of airport details
}
\description{
scrapWikipediaAirportsDetails
}
\examples{
scrapWikipediaAirportsDetails(url = "https://en.wikipedia.org/wiki/Bahrain_International_Airport")
}
