% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-wikipedia-airports-table.R
\name{scrapWikipediaAirportsTable}
\alias{scrapWikipediaAirportsTable}
\title{scrapWikipediaAirportsTable}
\usage{
scrapWikipediaAirportsTable(url)
}
\arguments{
\item{url}{a wikipedia airport table url}
}
\value{
a data.table with IATA, ICAO, Airport name, location served, time
DST and wikipedia url
}
\description{
get airport table data from wikipedia
}
\examples{
scrapWikipediaAirportsTable(
   url = 'https://en.wikipedia.org/wiki/List_of_airports_by_IATA_code:_A'
   )
}
\seealso{
https://en.wikipedia.org/wiki/List_of_airports_by_IATA_code:_B
}
