% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-vector-names.R
\name{setVectorName}
\alias{setVectorName}
\title{setVectorName}
\usage{
setVectorName(x = NULL, names = NULL)
}
\arguments{
\item{x}{a list}

\item{names}{names of the list}
}
\value{
a named list
}
\description{
setVectorName
}
\examples{
setVectorName(x = list(1,2,3,4), names = c("a","b","c","d"))
}
