% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{stationary.migration.summary}
\alias{stationary.migration.summary}
\title{find potential stationary periods and estimates their location and movement schedule}
\usage{
stationary.migration.summary(Result, prob.cutoff = 0.1, min.stay = 3)
}
\arguments{
\item{Result}{FLightR result object obtained from \code{\link{run.particle.filter}}}

\item{prob.cutoff}{Minimum probability that defines movement}

\item{min.stay}{Minimum duration of stationary period (in twilights)}
}
\value{
list with stationary and movement statistics
}
\description{
This function will find any sites where birds stayed longer than \code{min.stay}. Potential movement is detected by the minimum probability of movement \code{prob.cutoff}.
}
\author{
Eldar Rakhimberdiev
}

