% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{FLightRMovebank}
\alias{FLightRMovebank}
\title{Summary of estimated locations for Movebank}
\usage{
FLightRMovebank(Result, alpha = 0.5, filename = NULL)
}
\arguments{
\item{Result}{FLightR result object obtained from \code{\link{run.particle.filter}}
\code{estelleMetropolis} or \code{stellaMetropolis}.}

\item{alpha}{coverage of the credible intervals for now only two options: 0.95 or 0.5.}
}
\value{
either csv object or NULL if filename was specified
}
\description{
Summarize result object in the format required for upload to Movebank
}
\details{
This function accepts FLightR results object.
}
\author{
Eldar Rakhimberdiev
}
