% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ports2TAGS.R
\name{GeoLight2TAGS}
\alias{GeoLight2TAGS}
\title{Function to write down twilights annotated in GeoLight package data in so-called tags format}
\usage{
GeoLight2TAGS(raw, gl_twl, filename = NULL)
}
\arguments{
\item{raw}{original data - dataframe with two columns first column must contain time and second measured light levels}

\item{gl_twl}{twilights object from GeoLight}

\item{filename}{if not NULL data.frame in TAGS format will be returned}
}
\value{
\code{NULL} if \code{filename} is provided or TAGS formatted dataframe.
}
\description{
this function converts combines twilights detected in BAStag ot twGeos with raw data and writes them down in TAGS format that can be easily read by \code{\link{get.tags.data}}
}
\author{
Eldar Rakhimberdiev & Simeon Lisovski
}

