% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{make.prerun.object}
\alias{make.prerun.object}
\title{combines data, calibration and sets up priors}
\usage{
make.prerun.object(Proc.data, Grid, start, end = start, Calibration,
  threads = -1, Decision = 0.1, Direction = 0, Kappa = 0,
  M.mean = 300, M.sd = 500, likelihood.correction = TRUE)
}
\arguments{
\item{Proc.data}{Processed dat aobject created by \code{\link{get.tags.data}}.}

\item{Grid}{Spatial grid created by \code{\link{make.grid}}.}

\item{start}{release location (lat, lon).}

\item{end}{end of the track location. Will use \code{start} by default. Use NA in case of unknown end point.}

\item{Calibration}{Calibration object created by \code{\link{make.calibration}}.}

\item{threads}{number of parallel threads to use. default is -1, which means FLightR will use all available threads except 1.}

\item{Decision}{prior for migration probability values from 0 to 1 are allowed}

\item{Direction}{Direction prior for direction of migration (in degrees) with 0 pointing to the North}

\item{Kappa}{concentration parameter for vonMises distribution, 0 means uniform or even distribution. Will set some prioir for direction for all the track, so is not recommended to be changed}

\item{M.mean}{Prior for mean distance travelled between consequtive twilights, km}

\item{M.sd}{Prior for sd of distance travelled between consequtive twilights, the higher the value is the wider is the the distribution}
}
\value{
Object to be uses in the \code{\link{run.particle.filter}}
}
\description{
This function is one step before \code{\link{run.particle.filter}}. It combines data, calibration, spatial extent and movement priors and estimates spatial likelihoods that used later in the particle filter.
}
\author{
Eldar Rakhimberdiev
}

