% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_metrics.R
\name{match_gps_to_twilights}
\alias{match_gps_to_twilights}
\title{Match GPS fixes to FLightR twilight times}
\usage{
match_gps_to_twilights(Result, gps, max_time_diff_hours = 6)
}
\arguments{
\item{Result}{A FLightR Result-like object containing twilight times.}

\item{gps}{A data.frame with time, lon, and lat columns.}

\item{max_time_diff_hours}{Maximum allowed absolute time difference.}
}
\value{
A data.frame of matched twilight and GPS records.
}
\description{
Match GPS fixes to FLightR twilight times
}
