% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_particle_filter.R
\name{run.particle.filter}
\alias{run.particle.filter}
\title{Run Particle Filter}
\usage{
run.particle.filter(all.out, cpus = NULL, threads = -1,
  nParticles = 1e+06, known.last = T, precision.sd = 25,
  behav.mask.low.value = 0, save.memory = T, k = NA, parallel = T,
  plot = T, prefix = "pf", extend.prefix = T, max.kappa = 100,
  min.SD = 25, cluster.type = "PSOCK", a = 45, b = 1500, L = 90,
  adaptive.resampling = 0.99, check.outliers = F, sink2file = F,
  add.jitter = FALSE)
}
\arguments{
\item{all.out}{An object created by \code{\link{make.prerun.object}}.}

\item{threads}{An amount of threads to use while running in parallel. default is -1.}
}
\value{
FLightR object, containing output and extracted results. It is a list with the following elements 
\describe{
   \item{Indices}{List with prior infomration and indices}
   \item{Spatial}{Spatial data - Grid, Mask, spatial likelihood}
   \item{Calibration}{all calibration parameters}
   \item{Data}{original data}
   \item{Results}{The main results object. Main components of it are
      \describe{ 
      \item{Quantiles} {dataframe containing results on locations. Each line corresponds to a twilight}
      \item{Movement.results} {dataframe containing all the movement results, Note - time at line n means time of the end of transition between n and n-1}
      \item{outliers} {id of twilights excluded by online outlier detection tool}
      \item{LL} {negative loglikelihood}
      \item{Points.rle} {run length encoding object with posterior distribution for every twilight. Note that numbers of points correspond to line numbers in \code{$Spatial$Grid}}
      \item{Transitions.rle} {run length encoding object with all the transitions}
       }
  }
  }
}
\description{
\code{run.particle.filter}
}
\details{
This is the main function of FLightR, it takes fully prepared object created by \code{\link{make.prerun.object}} nd produces a result object that can be used for plotiing.
}

